src-py/pyproject.toml
[tool.poetry]
name = "dreamcanvas-backend"
version = "0.1.0"
description = "DreamCanvas FastAPI 服务"
authors = ["DreamCanvas Team <dev@dreamcanvas.local>"]
license = "MIT"
packages = [{ include = "dreamcanvas" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = { version = "^0.30.0", extras = ["standard"] }
httpx = "^0.27.0"
pydantic = "^2.7.1"
pydantic-settings = "^2.4.0"
python-multipart = "^0.0.9"
tenacity = "^8.3.0"
cryptography = "^43.0.0"
pillow = "^10.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.6"
httpx = "^0.27.0"
ruff = "^0.4.6"
mypy = "^1.10.0"

[tool.poetry.scripts]
dc-cli = "dreamcanvas.cli:main"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
